Make /licenses work better without JS, etc.
Preload /oss-licenses.json to make it load faster.

The preload strategy for the font I tried was not working as intended
(it also preloaded it when not referenced) so I switched back to
specifying <link rel="preload"> and deleted hooks.server.ts.

Deleted the leftover Makefile and unused webp (we now handle this
dynamically during build.)